Data/config_app/requirements.txt
psycopg2-binary
requests
confluent_kafka
asyncpg
asyncio
faust-streaming[prometheus]
kafka-python
aiokafka==0.8.1